+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 4.1.10 So Far... [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
 
 - "Must follow block" martial arts response element no longer behaves as a block maneuver. [#2528](https://github.com/dmdorman/hero6e-foundryvtt/issues/2528)
-- Added "Toggle Hold" and "Toggle Abort" to Combat Tracker context menu.
+- Added "Toggle Hold" and "Toggle Abort" to Combat Tracker context menu. [#2563](https://github.com/dmdorman/hero6e-foundryvtt/issues/2563)
 - Players can now advance Combat Tracker when they are done with their turn. [#2569](https://github.com/dmdorman/hero6e-foundryvtt/issues/2569)
 
 ## Version 4.1.9 20250801